support for custom prometheus controllers
